requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# LLM
groq==0.15.0

# RAG
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1

# Document loading
pypdf2==3.0.1

# Tools
pandas==2.2.3

# Logging
loguru==0.7.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

langchain>=0.3
langchain-groq>=0.2
langgraph>=0.2